requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
tiktoken
python-dotenv
langsmith
